requirements.txt
asgiref==3.2.10
Django==3.1.2
django-multiselectfield==0.1.12
gunicorn==20.0.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
